+++ e17/libs/ewl/src/ewl_text.h
@@ -29,15 +29,14 @@ void ewl_text_set_font_size(Ewl_Widget * w, int size);
 int ewl_text_get_font_size(Ewl_Widget * w);
 void ewl_text_set_color(Ewl_Widget * w, int r, int g, int b, int a);
 void ewl_text_get_color(Ewl_Widget * w, int *r, int *g, int *b, int *a);
-void ewl_text_get_text_geometry(Ewl_Widget * w, double *xx, double *yy,
-				double *ww, double *hh);
+void ewl_text_get_text_geometry(Ewl_Widget * w, int *xx, int *yy,
+				int *ww, int *hh);
 void ewl_text_set_style(Ewl_Widget * w, char *s);
-void ewl_text_get_letter_geometry(Ewl_Widget * w, int index, double *xx,
-				  double *yy, double *ww, double *hh);
-void ewl_text_get_letter_geometry_at(Ewl_Widget * w, double x, double y,
-				     double *tx, double *ty, double *tw,
-				     double *th);
+void ewl_text_get_letter_geometry(Ewl_Widget * w, int index, int *xx,
+				  int *yy, int *ww, int *hh);
+void ewl_text_get_letter_geometry_at(Ewl_Widget * w, int x, int y,
+				     int *tx, int *ty, int *tw, int *th);
 void ewl_text_set_alignment(Ewl_Widget * w, Ewl_Alignment a);
-int ewl_text_get_index_at(Ewl_Widget * w, double x, double y, int *index);
+int ewl_text_get_index_at(Ewl_Widget * w, int x, int y, int *index);
 
 #endif
